core: Split ISettingsManager into Manager and Loader
I thiought about a way to use a SettingsManager in hpacker, and saw a
flaw in the design of settings manager. Currently, it is responsible for
loading initial settings from file. Probably, that was only temporary
measure, because that is not always useful, thus reducing flexibility
and bringing redundancy to the design.