stringutils: Return instead of reference in split & explode
(Damn, it was hard to describe that in 40 symbols)
I think, that returning vector out of function provides nicer syntax,
than passing reference to destionation object into same function.

It also has side effect of overwriting the destination object, which
probably is a wanted behavior more often than not (for example, I
recently had to use vector::clear() in a config parser).

That also means that it no loger returns number of received tokens,
however, I believe using vector::size() is more efficient, so that's not
a problem.

(Hmm, and why I've written all that? Nobody will ever read that)