stringutils: Optimize explode()
That is not really an optimizetion, just a thing that bothered me …
It always performed an operation at the end of last iteration of loop,
result of which was thrown away. That probably is changed by compilers,
however, that fact bothered me too much.

The other way it could've been done:

	size_t pos1 = 0;
	size_t pos2 = source.find(delim, pos1);
	holder.push_back(source.substr(pos1, pos2 - pos1));

	while(pos2 != std::string::npos) {
		pos1 = pos2 + delim.size();
		pos2 = source.find(delim, pos1);
		holder.push_back(source.substr(pos1, pos2 - pos1));
	}

I really hate to have duplicate lines of code, so I prefer slightly
less readable variant over this one.