hdf: Split Parser into Lexer and Parser
This is an attempt to fix an issue parsing "value = type:thing".
The problem lies within the fact that parser needs to look at next token
after `type`, and if it isn't a `:` (colon), then this isn't actually
type, but a `thing`. But to look ahead of current token, some changes
are needed, because token gets consumed and that results in an error.

I decided to split the parser, because it started getting too big.