Move ‘scene’ into ‘graphics’
There isn't really an universal scene representation, best way varies
from case to case (in some cases it is a flat list with separate spatial
index, in other cases a tree (graph) is better, and so on).
Scene shouldn't be part of graphics either, but since all of this is
part of Irrlicht wrapper, I'll move them out of the way to there.