graphics: Add program class
— Improve shader class
— Add program class

I decided to go for different design than I originally intended — it is
easier to implement and use (arguably), than class with "always
compiled" invariant. It is impossible to implement without also making
it non-moveable and non-copyable.

On the plus side, I think this design is more flexible, as it allows
re-linking (I am unsure about performacnce implications though)